TaskListFormもサイズを覚えて再起動時に再現するようにする
SizeFormManagerをFormSizeManagerとrenameし、あとは引数でどのformのサイズをsave,loadするのか指定できるようにしました。
