+++ ProjectsTM.UI.MainForm/ProjectsTM.UI.MainForm.csproj
@@ -126,7 +126,6 @@
     <Compile Include="SearchWorkitemForm.Designer.cs">
       <DependentUpon>SearchWorkitemForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="SizeInfoManager.cs" />
     <Compile Include="VersionForm.cs">
       <SubType>Form</SubType>
     </Compile>
