+++ ProjectsTM.UI.TaskList/TaskListForm.cs
@@ -1,5 +1,8 @@
 ﻿using ProjectsTM.Model;
+using ProjectsTM.Service;
 using ProjectsTM.ViewModel;
+using System;
+using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 
@@ -10,6 +13,8 @@ namespace ProjectsTM.UI.TaskList
         private readonly ViewData _viewData;
         private PatternHistory _history;
         private FormSize _formSize;
+        private static string AppConfigDir => Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "ProjectsTM");
+        private static string SizeInfoPath => Path.Combine(AppConfigDir, "FormSizeInfo.xml");
 
         public TaskListForm(ViewData viewData, PatternHistory patternHistory, FormSize formSize)
         {
@@ -20,9 +25,7 @@ namespace ProjectsTM.UI.TaskList
             this._formSize = formSize;
             gridControl1.ListUpdated += GridControl1_ListUpdated;
             gridControl1.Initialize(viewData, comboBoxPattern.Text, _formSize.TaskListColWidths, checkBoxShowMS.Checked);
-            var offset = gridControl1.GridWidth - gridControl1.Width;
-            this.Width += offset + gridControl1.VScrollBarWidth;
-            this.Height = formSize?.TaskListFormHeight > this.Height ? formSize.TaskListFormHeight : this.Height;
+            this.Size = FormSizeManager.Load(SizeInfoPath, "TaskListFormSize");
             this.FormClosed += TaskListForm_FormClosed;
             this.checkBoxShowMS.CheckedChanged += CheckBoxShowMS_CheckedChanged;
         }
@@ -58,6 +61,7 @@ namespace ProjectsTM.UI.TaskList
             {
                 _formSize.TaskListColWidths.Add(gridControl1.ColWidths[idx]);
             }
+            FormSizeManager.Save(Height, Width, SizeInfoPath, "TaskListFormSize");
         }
 
         public void Clear()
@@ -70,7 +74,6 @@ namespace ProjectsTM.UI.TaskList
             comboBoxPattern.Items.Clear();
             comboBoxPattern.Items.AddRange(_history.Items.ToArray());
         }
-
         private void buttonUpdate_Click(object sender, System.EventArgs e)
         {
             UpdateList();
